미니 프로젝트1 준비 완료

+++ src/main/java/shop/mtcoding/metamall/model/log/login/LoginLogRepository.java
@@ -2,5 +2,5 @@ package shop.mtcoding.metamall.model.log.login;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoginLogRepository extends JpaRepository<LoginLog, Integer> {
+public interface LoginLogRepository extends JpaRepository<LoginLog, Long> {
 }
